requirements.txt
fastapi
uvicorn
python-multipart
requests
pydantic
python-dotenv
sqlalchemy
passlib[bcrypt]
bcrypt==4.1.3
